@@ -22,19 +22,19 @@
     "sharp": "^0.33.4"
   },
   "devDependencies": {
-    "@prismicio/types": "^0.2.8",
-    "@types/node": "^18.19.28",
-    "@types/react": "^18.2.74",
-    "@types/react-dom": "^18.2.23",
-    "@typescript-eslint/parser": "^7.5.0",
+    "@prismicio/types": "^0.2.9",
+    "@types/node": "^20.14.9",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@typescript-eslint/parser": "^7.15.0",
     "eslint": "^8.57.0",
-    "eslint-config-next": "^14.1.4",
+    "eslint-config-next": "^14.2.4",
     "eslint-config-prettier": "^9.1.0",
-    "prettier": "^3.2.5",
+    "prettier": "^3.3.2",
     "schema-dts": "^1.1.2",
-    "stylelint": "^16.3.1",
+    "stylelint": "^16.6.1",
     "stylelint-config-css-modules": "^4.4.0",
-    "stylelint-config-standard": "^36.0.0",
-    "typescript": "^5.4.3"
+    "stylelint-config-standard": "^36.0.1",
+    "typescript": "^5.5.3"
   }
 }
