@@ -65,4 +65,4 @@ const nextConfig = {
   trailingSlash: true,
 }
 
-module.exports = nextConfig
+export default nextConfig
